(refactor) Project structure
