--- strategy-builder-gui/src/pages/Strategies/components/Index.vue
@@ -1,13 +1,0 @@
-<template>
-  <div> <Content /> </div>
-</template>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-import Content from './Content.vue';
-
-export default defineComponent({
-  components: { Content },
-});
-</script>
