+++ strategy-builder-gui/src/pages/Strategies/composables/useStrategiesFilter.ts
